DEV : use released versions of reqs
previously, we were installing some of the dependencies from
source, which could cause CI failues because of changes in the
dependencies.

for now, we only test against released version of the packages

Note that we should, in the future, go back to testing against the
source/master branch of dependencies, to catch errors early. But,
we should also be testing against released versions, and allowing
CI failues in the source versions of deps.

	modified:   travis-ci-requirements.txt